requirements.txt
streamlit==1.28.1
google-generativeai==0.8.3
chromadb==0.4.15
PyPDF2==3.0.1
langchain==0.0.348
langchain-google-genai==1.0.1
python-dotenv==1.0.0
tiktoken==0.5.2
sentence-transformers==2.2.2
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0